+++ app/models/dictionary.rb
@@ -1,3 +1,0 @@
-class Dictionary < ActiveRecord::Base
-  has_many :entries, :class_name => 'DictionaryEntry'
-end