+++ app/models/dictionary_entry.rb
@@ -1,4 +1,0 @@
-class DictionaryEntry < ActiveRecord::Base
-  belongs_to :dictionary
-  has_many :dictionary_references
-end