+++ app/models/dictionary_reference.rb
@@ -1,4 +1,0 @@
-class DictionaryReference < ActiveRecord::Base
-  belongs_to :lemma
-  belongs_to :dictionary_entry
-end